window fixes

+++ src/bltPictJpg.c
@@ -347,7 +347,7 @@ JpgSkipInputData(j_decompress_ptr commPtr, long numBytes)
             
             sprintf(mesg, "short buffer: wanted %lu bytes, bytes left is %lu",
                     (unsigned long)numBytes,
-                    Blt_DBuffer_Length(readerPtr->dBuffer));
+                    (unsigned long)Blt_DBuffer_Length(readerPtr->dBuffer));
             Tcl_DStringAppend(&errorPtr->ds, " ", -1);
             Tcl_DStringAppend(&errorPtr->ds, mesg, -1);
             ERREXIT(commPtr, 10);
